rename project to scout-apm-logging
