+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "scout-apm-python-logging"
+name = "scout-apm-logging"
 version = "0.1.0"
 description = "Scout APM Python Logging Agent"
 authors = ["Quinn Milionis <quinn@scoutapm.com>"]
