+++ scout_apm_logging/__init__.py
@@ -1,0 +1,1 @@
+from scout_apm_logging.handler import OtelScoutHandler
