--- scout_apm_python_logging/__init__.py
@@ -1,1 +1,0 @@
-from scout_apm_python_logging.handler import OtelScoutHandler
